+++ Source/Minerals/DynamicMineral.cs
@@ -411,10 +411,10 @@ namespace Minerals
         // The main function controlling what is done each time the map is looked at
         public void Look()
         {
-            var watch = System.Diagnostics.Stopwatch.StartNew();
+            //var watch = System.Diagnostics.Stopwatch.StartNew();
             SpawnDynamicMinerals();
-            watch.Stop();
-            Log.Message("SpawnDynamicMinerals() took: " + watch.ElapsedMilliseconds);
+            //watch.Stop();
+            //Log.Message("SpawnDynamicMinerals() took: " + watch.ElapsedMilliseconds);
         }
 
 
@@ -422,7 +422,7 @@ namespace Minerals
         {
             foreach (ThingDef_DynamicMineral mineralType in DefDatabase<ThingDef_DynamicMineral>.AllDefs)
             {
-                var watch = System.Diagnostics.Stopwatch.StartNew();
+                //var watch = System.Diagnostics.Stopwatch.StartNew();
                              
                 // Check that the map type is ok
                 if (! mineralType.CanSpawnInBiome(map))
@@ -445,7 +445,7 @@ namespace Minerals
                 }
                 
 
-                Log.Message("Trying to spawn " + mineralType.defName + " with prob of " + mineralType.spawnProb + " and " + numToCheck + " blocks");
+                //Log.Message("Trying to spawn " + mineralType.defName + " with prob of " + mineralType.spawnProb + " and " + numToCheck + " blocks");
 
 
                 // Try to spawn in a subset of positions
@@ -474,14 +474,14 @@ namespace Minerals
 
 
                     // Try to spawn at that location
-                    Log.Message("Trying to spawn " + mineralType.defName);
+                    //Log.Message("Trying to spawn " + mineralType.defName);
                     //mineralType.TrySpawnAt(aPos, map, 0.01f);
                     mineralType.SpawnCluster(map, aPos, Rand.Range(0.01f, 0.05f), Rand.Range(mineralType.minSpawnClusterSize, mineralType.maxSpawnClusterSize));
 
                 }
 
-                watch.Stop();
-                Log.Message("Spawning " + mineralType.defName + " took: " + watch.ElapsedMilliseconds);
+                //watch.Stop();
+                //Log.Message("Spawning " + mineralType.defName + " took: " + watch.ElapsedMilliseconds);
 
             }
         }