+++ Source/Minerals/initialization.cs
@@ -12,24 +12,20 @@ namespace Minerals
     public static class mapBuilder
     {
 
-        public static void initAll(Map map)
+        public static void initRocks(Map map)
         {
+            // Remove starting chunks
             if (MineralsMain.Settings.removeStartingChunksSetting)
             {
                 removeStartingChunks(map);
             }
-            initStaticMinerals(map);
-            Log.Message("Minerals loaded");
-        }
 
-        public static void initStaticMinerals(Map map)
-        {
             List<string> spawned =  new List<string>();
 
             // Spawn static minerals
             foreach (ThingDef_StaticMineral mineralType in DefDatabase<ThingDef_StaticMineral>.AllDefs)
             {
-                if (mineralType.GetType() == typeof(ThingDef_StaticMineral) && (! spawned.Contains(mineralType.defName)))
+                if (mineralType.newMapGenStep == "chunks" && mineralType.GetType() == typeof(ThingDef_StaticMineral) && (! spawned.Contains(mineralType.defName)))
                 {
                     mineralType.InitNewMap(map);
                     spawned.Add(mineralType.defName);
@@ -39,7 +35,7 @@ namespace Minerals
             // spawn dynamic minerals
             foreach (ThingDef_StaticMineral mineralType in DefDatabase<ThingDef_StaticMineral>.AllDefs)
             {
-                if (mineralType.GetType() == typeof(ThingDef_DynamicMineral) && (! spawned.Contains(mineralType.defName)))
+                if (mineralType.newMapGenStep == "chunks" && mineralType.GetType() == typeof(ThingDef_DynamicMineral) && (! spawned.Contains(mineralType.defName)))
                 {
                     mineralType.InitNewMap(map);
                     spawned.Add(mineralType.defName);
@@ -49,7 +45,7 @@ namespace Minerals
             // spawn large minerals
             foreach (ThingDef_StaticMineral mineralType in DefDatabase<ThingDef_StaticMineral>.AllDefs)
             {
-                if (mineralType.GetType() == typeof(ThingDef_BigMineral) && (! spawned.Contains(mineralType.defName)))
+                if (mineralType.newMapGenStep == "chunks" && mineralType.GetType() == typeof(ThingDef_BigMineral) && (! spawned.Contains(mineralType.defName)))
                 {
                     mineralType.InitNewMap(map);
                     spawned.Add(mineralType.defName);
@@ -59,16 +55,29 @@ namespace Minerals
             // spawn everything else
             foreach (ThingDef_StaticMineral mineralType in DefDatabase<ThingDef_StaticMineral>.AllDefs)
             {
-                if (! spawned.Contains(mineralType.defName))
+                if (mineralType.newMapGenStep == "chunks" && (! spawned.Contains(mineralType.defName)))
                 {
                     mineralType.InitNewMap(map);
                 }
             }
         }
 
+        public static void initIce(Map map)
+        {
+            foreach (ThingDef_StaticMineral mineralType in DefDatabase<ThingDef_StaticMineral>.AllDefs)
+            {
+                if (mineralType.newMapGenStep == "plants")
+                {
+                    mineralType.InitNewMap(map);
+                }
+            }
+            //Log.Message("Ice spawned");
+        }
+
+
         public static void removeStartingChunks(Map map)
         {
-            string[] toRemove = {"ChunkSandstone", "ChunkGranite", "ChunkLimestone", "ChunkSlate", "ChunkMarble", "ZF_ChunkBasalt", "ChunkClaystone", "Filth_RubbleRock", "AB_ChunkCragstone", "AB_ChunkMudstone", "AB_ChunkObsidian", "GU_ChunkRoseQuartz", "AB_ChunkSlimeStone"};
+            string[] toRemove = {"ChunkSandstone", "ChunkGranite", "ChunkLimestone", "ChunkSlate", "ChunkMarble", "ZF_ChunkBasalt", "ChunkClaystone", "Filth_RubbleRock", "AB_ChunkCragstone", "AB_ChunkMudstone", "AB_ChunkObsidian", "GU_ChunkRoseQuartz", "AB_ChunkSlimeStone", "ZF_ChunkMudstone"};
             List<Thing> thingsToCheck = map.listerThings.AllThings;
             for (int i = thingsToCheck.Count - 1; i >= 0; i--)
             {