+++ NEWS.md
@@ -1,7 +1,8 @@
-## Current
+## v1.3.1
 
 * Fixed mineral low spawn rate bug
-
+* Fixes dynamic mineral spawn bug
+* Added Musdtone
 
 ## v1.3.0